+++ packages/@headlessui-vue/src/hooks/use-id.ts
@@ -1,24 +1,22 @@
-import { inject, InjectionKey, provide } from 'vue'
+import * as Vue from 'vue'
 
-let GENERATE_ID: InjectionKey<() => string> = Symbol('headlessui.useid')
-
+let GENERATE_ID: Vue.InjectionKey<() => string> = Symbol('headlessui.useid')
 let globalId = 0
 
-export function useId() {
-  let generateId = inject(GENERATE_ID, () => {
-    return `${++globalId}`
-  })
+export const useId =
+  // Prefer Vue's `useId` if it's available.
+  // @ts-expect-error - `useId` doesn't exist in Vue < 3.5.
+  Vue.useId ??
+  function useId() {
+    let generateId = Vue.inject(GENERATE_ID, () => {
+      return `${++globalId}`
+    })
 
-  return generateId()
-}
-
+    return generateId()
+  }
 /**
- * This function allows users to provide a custom ID generator
- * as a workaround for the lack of stable SSR IDs in Vue 3.x.
- *
- * This Nuxt users use the Nuxt provided `useId` function
- * which is stable across SSR and client.
+ * This function allows users to provide a custom ID generator.
  */
 export function provideUseId(fn: () => string) {
-  provide(GENERATE_ID, fn)
+  Vue.provide(GENERATE_ID, fn)
 }